// models/countries.ts
export interface Country {
    name: Name;
    tld?: string[];
    cca2: string;
    ccn3?: string;
    cca3: string;
    independent?: boolean;
    status: string;
    unMember: boolean;
    currencies?: Currencies;
    idd: Idd;
    capital?: string[];
    altSpellings: string[];
    region: string;
    subregion?: string;
    languages?: Languages;
    translations: Translations;
    latlng: number[];
    landlocked: boolean;
    area: number;
    demonyms?: Demonyms;
    flag: string;
    maps: Maps;
    population: number;
    fifa?: string;
    car: Car;
    timezones: string[];
    continents: string[];
    flags: Flags;
    coatOfArms: CoatOfArms;
    startOfWeek: string;
    capitalInfo: CapitalInfo;
    postalCode?: PostalCode;
    cioc?: string;
    borders?: string[];
    gini?: Gini;
}

export interface Name {
    common: string;
    official: string;
    nativeName?: NativeName;
}

export interface NativeName {
    spa?: Spa;
    afr?: Afr;
    eng?: Eng;
    nbl?: Nbl;
    nso?: Nso;
    sot?: Sot;
    ssw?: Ssw;
    tsn?: Tsn;
    tso?: Tso;
    ven?: Ven;
    xho?: Xho;
    zul?: Zul;
    sin?: Sin;
    tam?: Tam;
    smo?: Smo;
    kor?: Kor;
    ita?: Ita;
    cnr?: Cnr;
    ces?: Ces;
    slk?: Slk;
    hun?: Hun;
    kat?: Kat;
    fra?: Fra;
    por?: Por;
    ara?: Ara;
    zdj?: Zdj;
    deu?: Deu;
    ltz?: Ltz;
    her?: Her;
    hgm?: Hgm;
    kwn?: Kwn;
    loz?: Loz;
    ndo?: Ndo;
    srp?: Srp;
    cat?: Cat;
    gsw?: Gsw;
    roh?: Roh;
    hin?: Hin;
    lit?: Lit;
    lat?: Lat;
    nfr?: Nfr;
    nrf?: Nrf;
    jam?: Jam;
    msa?: Msa;
    run?: Run;
    tvl?: Tvl;
    nld?: Nld;
    swa?: Swa;
    sag?: Sag;
    ell?: Ell;
    pap?: Pap;
    fin?: Fin;
    swe?: Swe;
    nya?: Nya;
    sqi?: Sqi;
    zho?: Zho;
    ton?: Ton;
    som?: Som;
    slv?: Slv;
    fij?: Fij;
    hif?: Hif;
    ron?: Ron;
    pol?: Pol;
    mfe?: Mfe;
    tkl?: Tkl;
    grn?: Grn;
    kaz?: Kaz;
    rus?: Rus;
    nno?: Nno;
    nob?: Nob;
    smi?: Smi;
    kal?: Kal;
    dan?: Dan;
    pov?: Pov;
    khm?: Khm;
    tir?: Tir;
    mah?: Mah;
    vie?: Vie;
    ben?: Ben;
    prs?: Prs;
    pus?: Pus;
    tuk?: Tuk;
    jpn?: Jpn;
    hye?: Hye;
    aym?: Aym;
    que?: Que;
    lao?: Lao;
    bjz?: Bjz;
    tet?: Tet;
    ukr?: Ukr;
    aze?: Aze;
    niu?: Niu;
    tgk?: Tgk;
    kin?: Kin;
    hrv?: Hrv;
    kir?: Kir;
    gle?: Gle;
    glv?: Glv;
    bos?: Bos;
    lav?: Lav;
    mya?: Mya;
    est?: Est;
    pih?: Pih;
    div?: Div;
    fas?: Fas;
    fil?: Fil;
    fao?: Fao;
    cal?: Cal;
    cha?: Cha;
    nep?: Nep;
    mri?: Mri;
    nzs?: Nzs;
    tur?: Tur;
    kon?: Kon;
    lin?: Lin;
    bis?: Bis;
    mlg?: Mlg;
    amh?: Amh;
    bul?: Bul;
    bar?: Bar;
    bwg?: Bwg;
    kck?: Kck;
    khi?: Khi;
    ndc?: Ndc;
    nde?: Nde;
    sna?: Sna;
    toi?: Toi;
    zib?: Zib;
    tha?: Tha;
    arc?: Arc;
    ckb?: Ckb;
    ind?: Ind;
    dzo?: Dzo;
    gil?: Gil;
    lua?: Lua;
    mkd?: Mkd;
    rar?: Rar;
    ber?: Ber;
    mlt?: Mlt;
    heb?: Heb;
    isl?: Isl;
    bel?: Bel;
    mey?: Mey;
    pau?: Pau;
    uzb?: Uzb;
    mon?: Mon;
    crs?: Crs;
    nor?: Nor;
    hat?: Hat;
    nau?: Nau;
    hmo?: Hmo;
    tpi?: Tpi;
    urd?: Urd;
}
// ...
export interface Currencies {
    NIO?: Nio;
    ZAR?: Zar;
    LKR?: Lkr;
    USD?: Usd;
    XCD?: Xcd;
    COP?: Cop;
    UYU?: Uyu;
    KRW?: Krw;
    EUR?: Eur;
    CZK?: Czk;
    HUF?: Huf;
    GEL?: Gel;
    XAF?: Xaf;
    PAB?: Pab;
    KMF?: Kmf;
    GBP?: Gbp;
    SHP?: Shp;
    JOD?: Jod;
    NAD?: Nad;
    RSD?: Rsd;
    DJF?: Djf;
    CHF?: Chf;
    INR?: Inr;
    SYP?: Syp;
    CVE?: Cve;
    GGP?: Ggp;
    JEP?: Jep;
    AED?: Aed;
    JMD?: Jmd;
    GNF?: Gnf;
    GMD?: Gmd;
    MYR?: Myr;
    BIF?: Bif;
    GIP?: Gip;
    XPF?: Xpf;
    XOF?: Xof;
    VES?: Ves;
    AUD?: Aud;
    TVD?: Tvd;
    KES?: Kes;
    AWG?: Awg;
    MWK?: Mwk;
    ALL?: All;
    BND?: Bnd;
    SGD?: Sgd;
    HKD?: Hkd;
    TOP?: Top;
    SOS?: Sos;
    BHD?: Bhd;
    NZD?: Nzd;
    SEK?: Sek;
    BBD?: Bbd;
    GYD?: Gyd;
    STN?: Stn;
    FJD?: Fjd;
    CNY?: Cny;
    MDL?: Mdl;
    PLN?: Pln;
    MUR?: Mur;
    PYG?: Pyg;
    KZT?: Kzt;
    NOK?: Nok;
    DKK?: Dkk;
    SSP?: Ssp;
    BRL?: Brl;
    KHR?: Khr;
    ERN?: Ern;
    ZMW?: Zmw;
    SBD?: Sbd;
    MOP?: Mop;
    VND?: Vnd;
    BSD?: Bsd;
    CAD?: Cad;
    BMD?: Bmd;
    BDT?: Bdt;
    QAR?: Qar;
    AFN?: Afn;
    JPY?: Jpy;
    AMD?: Amd;
    PEN?: Pen;
    LAK?: Lak;
    BZD?: Bzd;
    FKP?: Fkp;
    UAH?: Uah;
    AZN?: Azn;
    SDG?: Sdg;
    ARS?: Ars;
    SRD?: Srd;
    CLP?: Clp;
    TJS?: Tjs;
    TND?: Tnd;
    KYD?: Kyd;
    GHS?: Ghs;
    SLL?: Sll;
    RWF?: Rwf;
    LSL?: Lsl;
    KGS?: Kgs;
    IMP?: Imp;
    KWD?: Kwd;
    MRU?: Mru;
    BAM?: Bam;
    RUB?: Rub;
    ANG?: Ang;
    MMK?: Mmk;
    TMT?: Tmt;
    MXN?: Mxn;
    DZD?: Dzd;
    LYD?: Lyd;
    CUC?: Cuc;
    CUP?: Cup;
    MVR?: Mvr;
    IRR?: Irr;
    PHP?: Php;
    FOK?: Fok;
    NPR?: Npr;
    YER?: Yer;
    TRY?: Try;
    VUV?: Vuv;
    MGA?: Mga;
    RON?: Ron2;
    ETB?: Etb;
    BGN?: Bgn;
    TZS?: Tzs;
    DOP?: Dop;
    TTD?: Ttd;
    WST?: Wst;
    ZWL?: Zwl;
    THB?: Thb;
    IQD?: Iqd;
    KPW?: Kpw;
    EGP?: Egp;
    ILS?: Ils;
    IDR?: Idr;
    BTN?: Btn;
    SZL?: Szl;
    KID?: Kid;
    LRD?: Lrd;
    LBP?: Lbp;
    CDF?: Cdf;
    MKD?: Mkd2;
    CKD?: Ckd;
    MAD?: Mad;
    OMR?: Omr;
    ISK?: Isk;
    BYN?: Byn;
    AOA?: Aoa;
    GTQ?: Gtq;
    SAR?: Sar;
    BWP?: Bwp;
    UGX?: Ugx;
    PKR?: Pkr;
    MZN?: Mzn;
    PGK?: Pgk;
    HTG?: Htg;
    NGN?: Ngn;
    TWD?: Twd;
    CRC?: Crc;
    HNL?: Hnl;
    UZS?: Uzs;
    BOB?: Bob;
    MNT?: Mnt;
    SCR?: Scr;
}

export interface Nio {
    name: string;
    symbol: string;
}

export interface Zar {
    name: string;
    symbol: string;
}

export interface Lkr {
    name: string;
    symbol: string;
}

export interface Usd {
    name: string;
    symbol: string;
}

export interface Xcd {
    name: string;
    symbol: string;
}

export interface Cop {
    name: string;
    symbol: string;
}

export interface Uyu {
    name: string;
    symbol: string;
}

export interface Krw {
    name: string;
    symbol: string;
}

export interface Eur {
    name: string;
    symbol: string;
}

export interface Czk {
    name: string;
    symbol: string;
}

export interface Huf {
    name: string;
    symbol: string;
}

export interface Gel {
    name: string;
    symbol: string;
}

export interface Xaf {
    name: string;
    symbol: string;
}

export interface Pab {
    name: string;
    symbol: string;
}

export interface Kmf {
    name: string;
    symbol: string;
}

export interface Gbp {
    name: string;
    symbol: string;
}

export interface Shp {
    name: string;
    symbol: string;
}

export interface Jod {
    name: string;
    symbol: string;
}

export interface Nad {
    name: string;
    symbol: string;
}

export interface Rsd {
    name: string;
    symbol: string;
}

export interface Djf {
    name: string;
    symbol: string;
}

export interface Chf {
    name: string;
    symbol: string;
}

export interface Inr {
    name: string;
    symbol: string;
}

export interface Syp {
    name: string;
    symbol: string;
}

export interface Cve {
    name: string;
    symbol: string;
}

export interface Ggp {
    name: string;
    symbol: string;
}

export interface Jep {
    name: string;
    symbol: string;
}

export interface Aed {
    name: string;
    symbol: string;
}

export interface Jmd {
    name: string;
    symbol: string;
}

export interface Gnf {
    name: string;
    symbol: string;
}

export interface Gmd {
    name: string;
    symbol: string;
}

export interface Myr {
    name: string;
    symbol: string;
}

export interface Bif {
    name: string;
    symbol: string;
}

export interface Gip {
    name: string;
    symbol: string;
}
// ...
export interface Sdg {
    name: string;
}
// ...
export interface Idd {
    root?: string;
    suffixes?: string[];
}

export interface Languages {
    spa?: string;
    afr?: string;
    eng?: string;
    nbl?: string;
    nso?: string;
    sot?: string;
    ssw?: string;
    tsn?: string;
    tso?: string;
    ven?: string;
    xho?: string;
    zul?: string;
    sin?: string;
    tam?: string;
    smo?: string;
    kor?: string;
    ita?: string;
    cnr?: string;
    ces?: string;
    slk?: string;
    hun?: string;
    kat?: string;
    fra?: string;
    por?: string;
    ara?: string;
    zdj?: string;
    deu?: string;
    ltz?: string;
    her?: string;
    hgm?: string;
    kwn?: string;
    loz?: string;
    ndo?: string;
    srp?: string;
    cat?: string;
    gsw?: string;
    roh?: string;
    hin?: string;
    lit?: string;
    lat?: string;
    nfr?: string;
    nrf?: string;
    jam?: string;
    msa?: string;
    run?: string;
    tvl?: string;
    nld?: string;
    swa?: string;
    sag?: string;
    ell?: string;
    pap?: string;
    fin?: string;
    swe?: string;
    nya?: string;
    sqi?: string;
    zho?: string;
    ton?: string;
    som?: string;
    slv?: string;
    fij?: string;
    hif?: string;
    ron?: string;
    pol?: string;
    mfe?: string;
    tkl?: string;
    grn?: string;
    kaz?: string;
    rus?: string;
    nno?: string;
    nob?: string;
    smi?: string;
    kal?: string;
    dan?: string;
    pov?: string;
    khm?: string;
    tir?: string;
    mah?: string;
    vie?: string;
    ben?: string;
    prs?: string;
    pus?: string;
    tuk?: string;
    jpn?: string;
    hye?: string;
    aym?: string;
    que?: string;
    lao?: string;
    bjz?: string;
    tet?: string;
    ukr?: string;
    aze?: string;
    niu?: string;
    tgk?: string;
    kin?: string;
    hrv?: string;
    kir?: string;
    gle?: string;
    glv?: string;
    bos?: string;
    lav?: string;
    mya?: string;
    est?: string;
    pih?: string;
    div?: string;
    fas?: string;
    fil?: string;
    fao?: string;
    cal?: string;
    cha?: string;
    nep?: string;
    mri?: string;
    nzs?: string;
    tur?: string;
    kon?: string;
    lin?: string;
    bis?: string;
    mlg?: string;
    amh?: string;
    bul?: string;
    de?: string;
    bwg?: string;
    kck?: string;
    khi?: string;
    ndc?: string;
    nde?: string;
    sna?: string;
    toi?: string;
    zib?: string;
    tha?: string;
    arc?: string;
    ckb?: string;
    ind?: string;
    dzo?: string;
    gil?: string;
    lua?: string;
    mkd?: string;
    rar?: string;
    ber?: string;
    mlt?: string;
    heb?: string;
    isl?: string;
    bel?: string;
    mey?: string;
    pau?: string;
    uzb?: string;
    mon?: string;
    crs?: string;
    nor?: string;
    hat?: string;
    nau?: string;
    hmo?: string;
    tpi?: string;
    urd?: string;
}

export interface Translations {
    ara: Ara2;
    bre: Bre;
    ces: Ces2;
    cym: Cym;
    deu: Deu2;
    est: Est2;
    fin: Fin2;
    fra: Fra2;
    hrv?: Hrv2;
    hun: Hun2;
    ita: Ita2;
    jpn?: Jpn2;
    kor: Kor2;
    nld: Nld2;
    per?: Per;
    pol: Pol2;
    por: Por2;
    rus: Rus2;
    slk: Slk2;
    spa: Spa2;
    srp: Srp2;
    swe: Swe2;
    tur: Tur2;
    urd: Urd2;
    zho?: Zho2;
}
// ...
export interface Demonyms {
    eng: Eng2;
    fra?: Fra3;
}

export interface Eng2 {
    f: string;
    m: string;
}

export interface Fra3 {
    f: string;
    m: string;
}

export interface Maps {
    googleMaps: string;
    openStreetMaps: string;
}

export interface Car {
    signs?: string[];
    side: string;
}

export interface Flags {
    png: string;
    svg: string;
    alt?: string;
}

export interface CoatOfArms {
    png?: string;
    svg?: string;
}

export interface CapitalInfo {
    latlng?: number[];
}

export interface PostalCode {
    format: string;
    regex?: string;
}

export interface Gini {
    "2014"?: number;
    "2016"?: number;
    "2013"?: number;
    "2019"?: number;
    "2018"?: number;
    "2010"?: number;
    "2015"?: number;
    "2017"?: number;
    "2011"?: number;
    "2003"?: number;
    "2004"?: number;
    "2012"?: number;
    "2006"?: number;
    "2008"?: number;
    "1998"?: number;
    "2009"?: number;
    "1999"?: number;
    "2005"?: number;
    "1992"?: number;
}